[fix] Closing Tagbar window from another changed which window was the "previous"
* Issue:

When the current (winnr()) window is not the Tagbar one, the user can
trigger closing the Tagbar window.  The s:CloseWindow() function was
failing to keep the window that was deemed as the current's "previous"
as that, causing the surprising effect of not leaving things as they
were (from the current window's perspective) before triggering the
Tagbar window closing.

* Changes:

  * s:CloseWindow() ('autoload/tagbar.vim'):

    * Try to mark (if there is one) the "previous" and "current" windows
      using distinct "mark_id" values;

    * Try to jump to the "previous" (if such a marked window is found,
      which should be the case when there was a "previous" window number
      before closing the Tagbar one) window, and then jump to the marked
      "current" one.

* Status:

  * TODO: remove some "dead" (/"prev:") code/comments introduced by this
    commit;