[fix] Closing Tagbar window when "current" doesn't jump to the "previous" one
* Issue:

In the existing code, when the Tagbar window was the "current" one and
the "close" operation was triggered, a very strange attempt at jumping
to a window containing the buffer for which the Tagbar window was
associated was done.  This is at best counter-intuitive, as the user
would reasonably expect to return to the window that was deemed as the
Tagbar window's "previous" one.

* Changes:

  * s:CloseWindow() ('autoload/tagbar.vim'):

    * If there is a valid "previous" winnr, mark its window and jump to
      it after closing the Tagbar window.

    * Its "restorable" state (window height for a vertically-split
      Tagbar, for example) is restored after jumping to it.